requirements.txt
google-genai
GitPython